refactor: migrate project to TypeScript and apply  PR feedback
