--- src/main.jsx
@@ -1,7 +1,0 @@
-import { createRoot } from 'react-dom/client'
-import './index.css'
-import App from './App.jsx'
-
-createRoot(document.getElementById('root')).render(
-    <App />,
-)
