+++ src/main.tsx
@@ -1,0 +1,9 @@
+import { createRoot } from 'react-dom/client'
+import './index.css'
+import App from './App'
+
+const root = document.getElementById('root') as HTMLElement;
+
+createRoot(root).render(
+    <App />,
+)
