+++ src/services/api.service.ts
@@ -1,0 +1,10 @@
+import axios from "axios";
+
+const api=axios.create({
+    baseURL: 'https://sample/api/',
+    headers: {
+        "Content-Type": "application/json",
+      },
+})
+
+export default api;
